remove pickupSeriesHelper again
