--- src/services/api/pickupSeriesHelpers.js
@@ -1,58 +1,0 @@
-export function convert (entry) {
-  if (Array.isArray(entry)) {
-    return entry.map(convert)
-  }
-  else {
-    return {
-      ...entry,
-      startDate: new Date(entry.startDate),
-      rule: convertRule(entry.rule),
-    }
-  }
-}
-
-export function serialize (entry) {
-  if (entry.rule) {
-    return {
-      ...entry,
-      rule: serializeRule(entry.rule),
-    }
-  }
-  else {
-    return entry
-  }
-}
-
-export function convertRule (rule) {
-  // defaults
-  let obj = {
-    byDay: ['MO'],
-    freq: 'WEEKLY',
-    isCustom: false,
-    custom: rule,
-  }
-  // takes rule string and returns object
-  let parts = rule.split(';')
-  if (parts.length > 2 || !rule.includes('FREQ=WEEKLY;BYDAY=')) {
-    return {
-      ...obj,
-      isCustom: true,
-    }
-  }
-  for (let part of parts) {
-    if (part.substr(0, 5) === 'BYDAY') {
-      obj.byDay = part.substr(6).split(',')
-    }
-    else if (part.substr(0, 4) === 'FREQ') {
-      obj.freq = part.substr(5)
-    }
-  }
-  return obj
-}
-
-export function serializeRule (obj) {
-  if (obj.isCustom) return obj.custom
-
-  // takes rule object and return string
-  return `FREQ=${obj.freq};BYDAY=${obj.byDay.join()}`
-}
