+++ src/socket.js
@@ -9,7 +9,7 @@ import { camelizeKeys } from '@/services/utils'
 import { convertDate as convertMessage } from '@/services/api/messages'
 import { parseDates as convertConversation } from '@/services/api/conversations'
 import { convertDate as convertPickup } from '@/services/api/pickups'
-import { convert as convertSeries } from '@/services/api/pickupSeriesHelpers'
+import { convert as convertSeries } from '@/services/api/pickupSeries'
 import { parse as convertFeedback } from '@/services/api/feedback'
 import { parseImageURLs as convertAuthUser } from '@/services/api/authUser'
 import { convertDates as convertHistory } from '@/services/api/history'
